requirements.txt
decorator==4.0.4
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-extensions==1.5.9
django-rest-swagger==0.3.4
django-toolbelt==0.0.1
djangorestframework==3.3.1
gunicorn==19.4.1
heroku==0.1.4
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
MySQL-python==1.2.5
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
psycopg2==2.6.1
ptyprocess==0.5
pygraphviz==1.3.1
python-dateutil==1.5
PyYAML==3.11
requests==2.8.1
simplegeneric==0.8.1
six==1.10.0
static3==0.6.1
traitlets==4.0.0
wheel==0.24.0
